+++ src/main/java/com/aorri2/goodsforyou/user/application/config/NewUserPolicyConfig.java
@@ -1,12 +1,12 @@
-package com.aorri2.goodsforyou.user.domain.config;
+package com.aorri2.goodsforyou.user.application.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import com.aorri2.goodsforyou.user.application.policy.NewUserEmailPolicy;
+import com.aorri2.goodsforyou.user.application.policy.NewUserNamePolicy;
+import com.aorri2.goodsforyou.user.application.policy.NewUserPasswordPolicy;
 import com.aorri2.goodsforyou.user.domain.UserFinder;
-import com.aorri2.goodsforyou.user.domain.policy.NewUserEmailPolicy;
-import com.aorri2.goodsforyou.user.domain.policy.NewUserNamePolicy;
-import com.aorri2.goodsforyou.user.domain.policy.NewUserPasswordPolicy;
 
 @Configuration
 public class NewUserPolicyConfig {
